+++ app/src/main/java/com/chs/your_body_profile/domain/usecase/GetDayMealTypeFatUseCase.kt
@@ -13,7 +13,7 @@ class GetDayMealTypeFatUseCase @Inject constructor(
     operator fun invoke(
         localDate: LocalDate,
         mealType: MealType
-    ): Flow<Double> {
+    ): Flow<Float> {
         return repository.getMealTypeTotalFat(localDate, mealType)
     }
 }
