+++ app/src/main/java/com/chs/your_body_profile/domain/usecase/GetDayMealTypeProteinUseCase.kt
@@ -12,7 +12,7 @@ class GetDayMealTypeProteinUseCase @Inject constructor(
     operator fun invoke(
         localDate: LocalDate,
         mealType: MealType
-    ): Flow<Double> {
+    ): Flow<Float> {
         return repository.getMealTypeTotalProtein(localDate, mealType)
     }
 }
